more

+++ wrapped-itoa/src/decls/macro_21.rs
@@ -1,6 +1,6 @@
 macro_rules! macro_21 {
     () => {
-        impl_Integer_size ! (isize as i32 # [cfg (target_pointer_width = "32")]) ;
+        impl_Integer_size ! (isize as i16 # [cfg (target_pointer_width = "16")]) ;
     };
 }
