+++ wrapped-itoa/src/decls/macro_22.rs
@@ -1,6 +1,6 @@
 macro_rules! macro_22 {
     () => {
-        impl_Integer_size ! (usize as u32 # [cfg (target_pointer_width = "32")]) ;
+        impl_Integer_size ! (usize as u16 # [cfg (target_pointer_width = "16")]) ;
     };
 }
