+++ wrapped-itoa/src/decls/macro_23.rs
@@ -1,6 +1,6 @@
 macro_rules! macro_23 {
     () => {
-        impl_Integer_size ! (isize as i64 # [cfg (target_pointer_width = "64")]) ;
+        impl_Integer_size ! (isize as i32 # [cfg (target_pointer_width = "32")]) ;
     };
 }
