+++ wrapped-itoa/src/decls/macro_24.rs
@@ -1,6 +1,6 @@
 macro_rules! macro_24 {
     () => {
-        impl_Integer_size ! (usize as u64 # [cfg (target_pointer_width = "64")]) ;
+        impl_Integer_size ! (usize as u32 # [cfg (target_pointer_width = "32")]) ;
     };
 }
